Unblock user input
Events are now separated between matrix events and the ones that affect
the user experience (input and window resizing). The events are pushed
to separate queues and  event loop now randomnly selects between those.
It might be better to always give precedence to user input, but this is
simpler and seems to work fine so far.